+++ src/components/accordion-header/accordion-header.tsx
@@ -2,7 +2,7 @@ import { Component, h, Host } from '@stencil/core';
 
 @Component({
     tag: 'wcs-accordion-header',
-    // styleUrl: 'accordion-header.scss',
+    styleUrl: 'accordion-header.scss',
     shadow: true
 })
 export class AccordionHeader {